requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
ormar==0.12.1
psycopg2-binary==2.9.5
alembic==1.13.2
SQLAlchemy==1.4.41
pydantic[dotenv]
pytest==7.3.1
pytest-asyncio==0.23.8
pytest-cov==6.0.0
ipython==8.31.0
python-dotenv==1.0.0
asyncpg==0.27.0
aiosqlite==0.17.0